requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
appdirs==1.4.4
Appium-Python-Client==2.6.1
APScheduler==3.10.4
apturl==0.5.2
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==21.11.1
Automat==20.2.0
backcall==0.2.0
base58==1.0.3
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
blinker==1.4
blis==0.7.11
boto3==1.24.57
botocore==1.27.57
Brlapi==0.8.3
Brotli==1.0.9
browsermob-proxy==0.8.0
build==0.10.0
cachelib==0.10.2
cachetools==5.3.2
catalogue==2.0.10
cbor==1.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cliapp==1.20180812.1
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cmdtest==0.32+git
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
confection==0.1.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==37.0.4
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.8
Cython==0.29.28
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
deprecation==2.1.0
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.18.0b1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.12.2
Fiona==1.9.4.post1
Flask==2.2.2
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-WTF==1.2.1
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.38.0
fs==2.4.12
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.5.2
GDAL==3.4.1
geographiclib==1.52
geopandas==0.13.2
geopy==2.2.0
google-api-core==2.12.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
imageio==2.22.0
imageio-ffmpeg==0.4.7
img2pdf==0.4.4
importlib-metadata==4.6.4
imutils==0.5.4
incremental==21.3.0
ipython==7.31.1
itsdangerous==2.1.2
jaro-winkler==2.0.0
jedi==0.18.0
jeepney==0.7.1
jellyfish==1.0.3
Jinja2==3.1.2
jmespath==1.0.1
kaitaistruct==0.10
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libusb1==3.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
M2Crypto==0.38.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.3
mnemonic==0.19
mongoengine==0.24.2
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpi4py==3.1.3
mpmath==0.0.0
msgpack==1.0.3
murmurhash==1.0.10
mysql-connector-python==8.2.0
netifaces==0.11.0
networkx==3.1
notify2==0.3
numpy==1.23.3
oauthlib==3.2.0
ocrmypdf==14.3.0
olefile==0.46
opencv-python==4.7.0.68
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.4.4
paramiko==2.9.3
parso==0.8.1
passlib==1.7.4
pbkdf2==1.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.2.1
Pillow==9.0.1
pillow-avif-plugin==1.4.1
pip-tools==6.12.2
pipreqs==0.4.11
platformdirs==3.9.1
pluggy==1.2.0
ply==3.11
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.28
proto-plus==1.22.3
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pure-python-adb==0.3.0.dev0
py-ubjson==0.16.1
pyasn1==0.5.0
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyjarowinkler==1.8
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF2==2.11.1
pypng==0.0.20
pyproj==3.4.0
pyproject_hooks==1.0.0
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.20.0
python-snappy==0.5.3
pythran==0.10.0
PyTrie==0.4.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.1
reportlab==3.6.8
requests==2.31.0
rsa==4.9
rsa-python==0.1.1
s3transfer==0.6.0
schedule==1.1.0
scikit-image==0.21.0
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.3.0
selenium-wire==5.1.0
service-identity==18.1.0
shapely==2.0.1
six==1.16.0
smart-open==6.4.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speedtest-cli==2.1.3
sqlparse==0.4.2
srsly==2.4.8
ssh-import-id==5.11
sympy==1.9
systemd-python==234
terminator==2.1.1
tesserocr==2.6.1
textdistance==4.6.0
thinc==8.2.1
tifffile==2023.7.18
timeout-decorator==0.5.0
tomli==2.0.1
tqdm==4.64.1
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
ttystatus==0.38
twilio==7.12.1
Twisted==22.1.0
txaio==21.2.1
typer==0.9.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2022.2
tzlocal==4.2
u-msgpack-python==2.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.0
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
Werkzeug==2.2.2
wifi==0.3.8
wsaccel==0.6.3
wsproto==1.1.0
WTForms==3.1.1
xdg==5
xkit==0.0.0
XlsxWriter==3.0.3
yarg==0.1.9
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.22.0
